backend/requirements.txt
# Web Framework
fastapi
uvicorn

# AI & LLM
openai
azure-ai-inference
tiktoken

# Document Processing
pypdf
python-docx
pdfplumber

# Database
sqlalchemy
psycopg2-binary
alembic

# Utilities
pydantic
pytest
python-dotenv

# File Upload
python-multipart

# Password Hashing
passlib
bcrypt

# Authentication
fastapi-login


